// src/app/Apps/Agispay/model/agismodel.ts
export interface AgisModel {
  nameOfPlotOwner: string;
  occupation: string;
  idType: string;
  cadestralZone: string;
  emailOfPlotOwner: string;
  dateOfBirth: string;
  address: string;
  idNumber:string;
  plotNumber: string;
  amount: number;
  paymentDescription: string;
  depositorPhoneNumber: string;
  fileNumber: string;
  paymentFor: string;
  depositorSlip: string;
  depositorName: string;
  plotOwnerPhoneNumber: string;
}